+++ libft/ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: sabrown <sabrown@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/29 21:41:25 by sabrown           #+#    #+#             */
-/*   Updated: 2025/05/12 00:47:18 by sabrown          ###   ########.fr       */
+/*   Updated: 2025/05/15 01:17:42 by sabrown          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t    ft_strlcpy(char *dest, const char *src, size_t size)
 {
-    size_t     l;
+    int     l;
     size_t     i;
 
     l = ft_strlen(src);
